# docs/swagger.yaml
basePath: /
host: localhost:5000
info:
  contact:
    name: API Support
  description: swagger Projeto Raul
  title: Fiber Example API
  version: "1.0"
paths:
...
  /deleteEnvironment/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Environment by its ID
      operationId: delete-environment-by-int
      parameters:
      - description: Environment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Environment
      tags:
      - Environment
  /deleteItem/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Item by its ID
      operationId: delete-item-by-int
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Item
      tags:
      - Item
  /deleteLandscape/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Landscape by its ID
      operationId: delete-landscape-by-int
      parameters:
      - description: Landscape ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Landscape
      tags:
      - Landscape
  /deletePlayer/{id}:
    delete:
      consumes:
      - application/json
      description: delete a Player by its ID
      operationId: delete-reature-by-int
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
      summary: delete a Player
      tags:
      - Player
...
  /environment:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Environment
  /environments:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Environment
  /item:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Item
  /items:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Item
  /landscape:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Landscape
  /landscapes:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Landscape
  /player:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Insert new Player
  /players:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: Get  list of Player
...
  /updateEnvironment/{id}:
    put:
      consumes:
      - application/json
      description: update a Environment by its ID
      operationId: get-Environment-by-int
      parameters:
      - description: Environment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Environment
      tags:
      - Item
  /updateItem/{id}:
    put:
      consumes:
      - application/json
      description: update a Item by its ID
      operationId: get-Item-by-int
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Item
      tags:
      - Item
  /updateLandscape/{id}:
    put:
      consumes:
      - application/json
      description: update a Landscape by its ID
      operationId: get-Landscape-by-int
      parameters:
      - description: Landscape ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Landscape
      tags:
      - Landscape
  /updatePlayer/{id}:
    put:
      consumes:
      - application/json
      description: update a Player by its ID
      operationId: get-Player-by-int
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
      summary: update a Player
      tags:
      - Player
...
swagger: "2.0"